stats: pretty much rewrite it
bots.discordlist.net is no longer a thing, and the API for PW and DBL
is exactly the same. This means we can just use one cog! The best part is,
since there's only one cog, the stats command only gets registered once,
which means that this cog can be reloaded without errors.